backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
supabase==2.11.0
python-magic==0.4.27
Pillow==11.1.0
cairosvg==2.7.1
pdf2image==1.17.0
pdf2docx==0.5.8
python-docx==1.1.2
fpdf2==2.8.2
